+++ database/migrations/2025_08_24_224902_0005_add_image_to_events_table.php
@@ -12,7 +12,9 @@ return new class extends Migration
     public function up(): void
     {
         Schema::table('events', function (Blueprint $table) {
-             $table->string('image')->nullable()->after('description');
+            if (!Schema::hasColumn('events', 'image')) {
+                $table->string('image')->nullable()->after('description');
+            }
         });
     }
 
@@ -22,7 +24,9 @@ return new class extends Migration
     public function down(): void
     {
         Schema::table('events', function (Blueprint $table) {
-            $table->dropColumn('image');
+            if (Schema::hasColumn('events', 'image')) {
+                $table->dropColumn('image');
+            }
         });
     }
 };
